backend/requirements.txt
pyzmq==25.1.1
duckdb==0.9.2
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1
lxml==4.9.3
pdfplumber==0.10.3
cryptography==41.0.7
bcrypt==4.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
loguru==0.7.2
tqdm==4.66.1
requests==2.31.0
